all: cleanup unused code, part 2
And enable U1000 check in staticcheck.

Updates #cleanup
