+++ staticcheck.conf
@@ -14,4 +14,5 @@ checks = [
 
 	"QF1004", # Use `strings.ReplaceAll` instead of `strings.Replace` with `n == 1`
 	"QF1006", # Lift if+break into loop condition
+	"U1000", # catch unused code
 ]
