+++ tstest/integration/vms/opensuse_leap_15_1_test.go
@@ -1,6 +1,8 @@
 // Copyright (c) Tailscale Inc & AUTHORS
 // SPDX-License-Identifier: BSD-3-Clause
 
+//go:build !windows && !plan9
+
 package vms
 
 import (
